Moving PdfLatexFiles into own class.
