+++ src/main/java/github/paz/awardportal/pdf/PdfLatexCommandRunner.java
@@ -53,39 +53,4 @@ public class PdfLatexCommandRunner {
         String arg = COMMAND.format(templateFileName);
         return Arrays.asList("bash", "-c", arg);
     }
-
-    /*
-    Expected output files generated by pdflatex against a given template file.
-     */
-    @Getter
-    private class PdfLatexFiles {
-        private Path pdfFile;
-        private Path auxFile;
-        private Path logFile;
-        // Other byproducts?
-
-        public PdfLatexFiles(Path texTemplate) {
-            String texTemplateFileName = getFileName(texTemplate);
-            String pdfFileName = changeExtension(texTemplateFileName, "pdf");
-            String auxFileName = changeExtension(texTemplateFileName, "aux");
-            String logFileName = changeExtension(texTemplateFileName, "log");
-
-            // TODO - need to worry about parent directories, not just file name?
-            pdfFile = Paths.get(pdfFileName);
-            auxFile = Paths.get(auxFileName);
-            logFile = Paths.get(logFileName);
-
-        }
-
-        private String getFileName(Path path) {
-            return path.getFileName().toString();
-        }
-
-        private String changeExtension(String fileName, String newExtension) {
-            int pos = fileName.lastIndexOf('.');
-            String noExtension = fileName.substring(0, pos);
-            return noExtension + '.' + newExtension;
-
-        }
-    }
 }
