+++ src/main/java/github/paz/awardportal/pdf/PdfLatexFiles.java
@@ -1,0 +1,49 @@
+package github.paz.awardportal.pdf;
+
+import lombok.Getter;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.List;
+
+/*
+Expected output files generated by pdflatex against a given template file.
+ */
+@Getter
+public class PdfLatexFiles {
+    private Path texTemplate;
+    private Path pdfFile;
+    private Path auxFile;
+    private Path logFile;
+    // Other byproducts?
+
+    public PdfLatexFiles(Path texTemplate) {
+        String texTemplateFileName = getFileName(texTemplate);
+        String pdfFileName = changeExtension(texTemplateFileName, "pdf");
+        String auxFileName = changeExtension(texTemplateFileName, "aux");
+        String logFileName = changeExtension(texTemplateFileName, "log");
+
+        // TODO - need to worry about parent directories, not just file name?
+        this.texTemplate = texTemplate;
+        pdfFile = Paths.get(pdfFileName);
+        auxFile = Paths.get(auxFileName);
+        logFile = Paths.get(logFileName);
+
+    }
+
+    public List<Path> getAll() {
+        return Arrays.asList(texTemplate, pdfFile, auxFile, logFile);
+    }
+
+    private String getFileName(Path path) {
+        return path.getFileName().toString();
+    }
+
+    private String changeExtension(String fileName, String newExtension) {
+        int pos = fileName.lastIndexOf('.');
+        String noExtension = fileName.substring(0, pos);
+        return noExtension + '.' + newExtension;
+
+    }
+}
